rename: composer directory is now named core
This change helps clarify the core of the tekton tool versus the components that help provide the inner details.
